requirements.txt
blinker==1.6.2
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.3
freezegun==1.2.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.2
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
Werkzeug==2.3.7
